Made spearate style sheets, removed changes to App.tsx
