+++ dfm-sideline-sidekick-app/components/BulletPointStyles.tsx
@@ -1,0 +1,40 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: "row",
+    alignItems: "center",
+    paddingTop: 15,
+  },
+  circle: {
+    width: 24, // this should be a "props"-value in future
+    height: 24, // this should be a "props"-value in future
+    borderRadius: 24 / 2,
+    backgroundColor: "#00629B",
+    alignItems: "center",
+    justifyContent: "center",
+  },
+
+  circleCaption: {
+    fontSize: 14,
+    color: "white",
+    fontWeight: "600",
+  },
+  mainText: {
+    color: "black",
+    fontSize: 16,
+    fontFamily: "Roboto",
+    fontWeight: "700",
+    paddingLeft: 5,
+  },
+  subpoints: {
+    paddingTop: 5,
+    paddingLeft: 38,
+    paddingRight: 25,
+  },
+  point: {
+    fontSize: 15,
+  },
+});
+
+export default styles;
