+++ dfm-sideline-sidekick-app/pages/generalPrinciplesStyles.tsx
@@ -1,0 +1,34 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: "#fff",
+    paddingTop: 50,
+    textAlign: "left",
+    alignSelf: "stretch",
+    paddingLeft: 25,
+  },
+  title: {
+    color: "#182B49",
+    fontSize: 28,
+    fontFamily: "Roboto",
+    fontWeight: "700",
+    marginBottom: 20,
+    textAlign: "left",
+    paddingTop: 10,
+  },
+  subTitle: {
+    color: "#182B49",
+    fontSize: 21,
+    fontFamily: "Roboto",
+    fontWeight: "700",
+    marginBottom: 5,
+    textAlign: "left",
+  },
+  button: {
+    fontSize: 25,
+  },
+});
+
+export default styles;
